requirements.txt
py-clob-client>=0.17.0
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
eth-account>=0.11.0
